Update syntax to create a WP_Widget object to avoid PHP deprecated warnings from Wordpress core
The use of $this->WP_Widget ends in a warning, because there is a WP_Widget method, that helps PHP 4 to call the contstuctor that is marked as deprecated since Wordpress 4.3.0. Hope this methode will be deleted soon! 
$this->WP_Widget would normally call the _constructor methode of the class, if there wasn't this PHP 4 helper method named exactly as the class. If the WP_Widget method will be deleted, switch back to $this->WP_Widget.
It's a Wordpress error, that I reported to them too, but please change to avoid problem til they changed it. Thx!